+++ app/src/main/res/navigation/nav_graph.xml
@@ -22,7 +22,7 @@
         tools:layout="@layout/fragment_repository">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.Item" />
+            app:argType="jp.co.yumemi.android.code_check.model.Item" />
     </fragment>
 
 </navigation>
